+++ opencascade-hs/src/OpenCascade/BRepAdaptor.hs
@@ -1,0 +1,5 @@
+module OpenCascade.BRepAdaptor
+(Curve
+) where
+
+import OpenCascade.BRepAdaptor.Types (Curve)